Add helper function setICUDataDirectory.
On android, icu4c is compiled without the data integrated to the library.
The data is a file that have to be loaded by icu at runtime.
It is up to the android application to configure icu to tell it where
to find the data.

The libraries are compiled as a dynamic library but linked to
its dependencies using static library (called here as mixed linking)
It means that functions from dependencies not used by the dynamic
library are stripped from the binary.

Until now, libkiwix android wrapper was integrated in libkiwix source
code and compiled with libkiwix. And it was wrapping libzim with the help
of the libkiwix wrappers (`Reader`, `Searcher`).
As it was compiled with libkiwix, the jni wrapper was calling
`u_setDataDirectory` and the symbol was included.

But with recent change this is not the case:
- Jni wrapper is moved in another repository.
- Libkiwix wrappers (`Reader`, ...) have been removed, meaning that jni
  wrapper needs to link with libzim directly.
- Both libkiwix and libzim are compiled using mixed linking
  (libkiwix still linking dynamically to libzim).

So it appears that none of libzim or libkiwix is using `u_setDataDirectory`
and the symbol is stripped. So the wrapper cannot use it.
We could distribute icu as a dynamic library and tell wrapper to link to
icu libraries too, but it seems a bit overkill
(especially as we would like to rethink how we link to icu and remove this
set data directory call in the future, hopefully).

By adding this small helper in libzim, `u_setDataDirectory` will not be
stripped and the wrapper will be able to set the data directory calling
`zim::setICUDataDirectory`.
The helper is added in libzim and not in libkiwix as libzim is the lower
level library which needs icu.