Enter Grouping<ObjectId, GroupId>
The new class Grouping<ObjectId, GroupId> implements the current
straightforward approach for sorting the dirents in cluster order, but
opens a way for switching to a better solution.

A user-observable effect of this change is the lower memory usage after
`FileImpl::prepareArticleListByCluster()` has completed (but an increased
high-watermark memory usage of that function).