Add tests showing how the creator should behave when buggy code is used.
Creator is based on API the user have to implement.
If the user create buggy code or throw exception, we must be prepared
to handle them.

The new tests show some use cases (all ?) we must be prepared to handle.
They are not handle now, will be in next commits.