Optimization of Entry::getItem()
For a client of libzim the only way to create an `Item` object is
through `Entry::getItem()` or `Entry::getRedirect()` (the latter
relying on the former to deliver the final result).

Before this change, `Entry::getItem()` performed (via the `Item`
constructor) an extra call to `FileImpl::getDirent(entry_index_t)`.
Now the result of that operation that is readily available in the source
`Entry` object is reused.