Remove duplicated line in test/error_in_creator.
Adding twice the same entry (metadata or not) will throw a exception.
But we are testing async error here.
If the async error not thrown in time because the writer thread is "slow",
we will not catch the async error in time and we will throw the "wrong"
exception.

Should fix #775