Fix ambiguous definition of operators for custom types.
Constraints for REAL_TYPE<B> are :
- Child of this class can be compared/incremented/...
- Return type of the operator must be the child and not `REAL_TYPE<B>`

The first point make us overloaded classic operators.

This second point was making us templating the operator by `T` and not
`REAL_TYPE<B>`.

But it is to generic as we end define `operator-` for any type with
`operator-=`.

So we need to overload the operators for `NAME` (in the macro) and not use
generics.