Avoid working on too long strings when removing accents.
My use case is a 20Mb french text string to index, on my computer,
a simple "transliterate" takes around 50 minutes to finish.

If we assume a ratio of 1/10 letter with accents, a 20 million char text
would need 20M/10[number of replacement] * 20M/2[number of char to move]
moves. It is something like : 2*10^13 char moves.

By using batch of 4K, we have 4K/10*4K/2=0.8M moves per batch.
And we have 20M/4K=5000 batches.
So a total of 5000*0.8M=4G (4*10^9).
The ratio is 5000. So we could expect a code running 5000 times faster.

In practice, it goes from 49 minutes to 3 or 4 seconds.