Rename `*Url*` symbols to `*Path*`.
Dirents' "url" are not really url. They are u8 arrays which serve
as the main "key".
Specification says that this u8 array store a utf-8 encoded string and
we use this key as a path.

Public API already use path semantic so rename all internal symbols
from url to path.

Fix #868