Make openSinglePieceOrSplitZimFile better handling single/multi part opening.
Now, it tries to open as single piece only if not ending by `.zimaa`.
If it cannot find the file, or if ending with `.zimaa`, it tries multi zim part.

In any case, FileImpl constructor check for size and throw a
ZimFileFormatError is file's size doesn't correspond to what is the header.