Do not try to read out of reader in release mode.
When building with the meson option `b_ndebug=true`
(never set on our CI/CD), `ASSERT` macro is removed and so we stop
testing that we are actually read in the reader range.

Now we introduce `readImpl` which do the actual read
(depending of implementation) and make `read` testing read range and call
`readImpl` when everything is ok.