Add a test on a broken archive with a invalid offset in cluster.
Not that the test is correctly catching invalid offset only for nons variant.
Current `validate` method is not checking the offsets in the cluster.
Nons variant detect the invalid offset as `titlePtrList` is stored in the cluster
and so we try to open it (using `X/listing/titleOrdered/vX`)

Correct test will be added in next commit.

Need PR https://github.com/openzim/zim-testing-suite/pull/9