Introduce MultiMutex, to lock multiple mutex in the same time
We will need to lock all db mutex in the same time when doing multizim
search. And we must avoid dead locks in the process.

Std lib provides us a `std::lock()` [1] function to lock several mutex
in the same time. However, it is a variatic function and it doesn't take
a list of mutex to lock.

So we havet to implemented ourselves. But we don't want to re implement
an error prone algorithm. To do so, we introduce a `MultiMutex` (a
linked list) and we (ab)use the fact that `std::lock()` is a template
taking any object "Lockable" [2]

[1] https://en.cppreference.com/w/cpp/thread/lock
[2] https://en.cppreference.com/w/cpp/named_req/Lockable