Move dirent_lookup configuration in the `OpenConfig`
Now m_direntLookup is created at fileimpl creation, not at first (user)
access to the dirents.

This is kind of useless to delay the dirent_lookup creation until the
first access as user will need to create it to use the archive.

For the rare case when this is not needed (search only, iter efficient),
user can now simply deactivate the dirent lookup.