Preloading the xapian database is now an explicit call.
We want to have xapian preloading optional.
However may want to do it out of the constructor, as we do for dirent
lookup. But, contrarly to dirent_lookup, we cannot wait the first access
to the db to load it as it would cancel all the purpose of preloading.

Using a new helper `preloadXapianDb` helps here.