Check if shared_future has been drop while we compute the value.
If creation of the value takes a lot of time, we may have move a lot of
value in front in the lru cache and drop the shared_future before we
set it.

So before blindly increase the cost, we check the promise is still here
and readd it else.