Backward compatibility of title index v0 removal
Absence of the title index v0 is now signaled via the titleIdxPos set to
0xffffffffffffffff instead of 0 since the latter resulted in ZIM files
incompatible with older versions of libzim. Several such ZIM files have
found home in zim-testing-suite (v 0.7.0) and have to be replaced. In
theory, each of those ZIM files could be patched with a few bytes in two
locations (titleIdxPos in the ZIM file header and the checksum at the
end of the file). An easier solution is to recreate them, but I don't
want it to be done with a locally built version of zim-tools (like it
was done in openzim/zim-testing-suite#13 which references a libzim
commit that was eventually edited/replaced during the review/rebase of
PR#949 and is no longer part of the main branch).  That's
`Fileheader::hasTitleListingV0()` considers the `titleIdxPos == 0` case
too. Once this PR is merged and we have an official CI build we can use
it to update zim-testing-suite, whereupon the `titleIdxPos == 0` test
can be dropped from `Fileheader::hasTitleListingV0()` (in a separate
PR).