Make caches open to a cost of each item different than 1
At cache level, "size" is renamed to "cost", representing how much a
cache can store (and how much an item cost). "size" is keep to count
the number of items.

At higher level, we keep the "size" sementics as we are speaking about
the size of the cache, whatever this is.

This is the first step to a cache limited by memory usage.