Fixed the race condition in NamedThread
Before this fix, `NamedThread::getCurrentThreadName()` might be called
from the new thread (that already started executing) before the
assignment to `thread_` in the `NamedThread` constructor had taken
place. Then the query to the thread id (`nt->thread_.get_id()`) on such
a thread did not return the correct id, and the end result was that a
new (numbered) thread name was synthesized for that thread. From that
point onward, concurrency orchestration via `log_debug()` (which relies
on correct thread names) broke, typically resulting in a deadlock.

This fix slightly reduces the generality of the `NamedThread` constructor
(support for arguments to the callable has been dropped) since it is not
currently needed (and even if it was needed, it could be circumvented
with an extra lambda expression).