A faster & more fragile Log.concurrencyOrchestration test
This reverts commit 3cf4e08 "Fix for the test failure under Macos CI"
which addressed the failure of the Log.concurrencyOrchestration unit
test in libzim CI by increasing the duration of the thread suspension
time windows intended to achieve the desired linearization of operations
in the two threads. Such an approach has proved to be unreliable in the
kiwix-build CI/CD.

Therefore the test is rolled back to its initial fragile state. The next
commit will provide another attempt to make it more stable.