Suspend the thread in multiple short shots
It seems that the unorchestrated execution of the 2 threaded flow in the
Log.concurrencyOrchestration unit test deviates from the intended
linearization under high load when the scheduler keeps a sleeping thread
in suspended state much longer than intended by the design of the test
flow. By suspending the thread in multiple brief shots instead of a
relatively long one we engage the scheduler more frequently and increase
the probability that it will pick the right thread.