Testing for handling of too long words in titles
Xapian's parser (TermGenerator) discards words/terms longer than a
certain limit (default: 64). However terms can be added to the indexed
documents directly via Xapian::Document::add_term(). We do that in order
to index titles that are made fully of non-word characters but our
implementation opens up a loophole for words of arbitrary length to slip
in (when the title is a single word). That leads to crashes if Xapian's
hard limit on the length of a term (max 245 characters) is exceeded.

The new unit tests demonstrates the existence of a loophole.