Protection against too long single-word titles
Titles consisting of a single word exceeding the max word length limit
are not indexed (similarly to how that long word would be handled if it
were not the only word in the title).

An alternative might be to index a truncated version of the title but
that would complicate the verification of the bug fix (because the
potentially dangerous title would remain in the expected output of the
unit test).