Made spelling suggestion case sensitive
... and fixed a bug in the test data.

This reduced the count of failures in the Suggestion.spellingSuggestions
unit test from 8 to 1:

[ RUN      ] Suggestion.spellingSuggestions
Resolve redirect
set index
test/suggestion.cpp:841: Failure
Expected equality of these values:
  getSpellingSuggestions(a, "Lax", 1)
    Which is: {}
  std::vector<std::string> ({"Lachs"})
    Which is: { "Lachs" }
[  FAILED  ] Suggestion.spellingSuggestions (260 ms)

The spelling correction "Lax -> Lachs" is not returned because the max
edit distance is capped at (length(query_word) - 1) which reduces our
passed value of the max edit distance argument from 3 to 2.

This problem disappears if the version of libxapian found on Ubuntu
22.04 (libxapian.so.30.11.0) is used instead of the one that we build
ourselves as a base dependency (libxapian.so.30.12.4).