Use public Xapian API
Xapian 1.4.12 and later have a Database::size() method which returns
the number of shards, so use this instead which fixes a compilation
failure with Xapian 2.0.x where the internal object is no longer a
std::vector.

This raises the Xapian version requirement to 1.4.12 (released 2019,
available in Ubuntu 20.04 and later), which is now checked by
meson.build (previously there wasn't a checked requirement
but at least 1.4.0 was needed for e.g. Xapian::LatLongCoord).