[HACK] Detection of omissions during title indexing
Xapian may silently omit words exceeding a certain size limit during
indexing and there is no API to detect that it happened. This commit
is a hacky way of detecting such omissions at the cost of false
positives trigerred by titles containing too much whitespace and/or
punctuation.

Also decreased the limit on the max term size because of crashes
in the newly extracted test-case Suggestion.handlingOfTooLongWords.